response: improve performance of seconds->gmt-bytes
`seconds->date' takes a second parameter controlling whether or not to
return a UTC date; this is much faster than generating a date in the
local timezone.

Instead of formatting a string and then converting it to bytes, we
write the result directly to a bytestring, avoiding allocations and
parsing of the format string.